more FILE_OFFSET_BITS == 64 adjustments.

+++ include/libbb.h
@@ -96,7 +96,7 @@ void reset_ino_dev_hashtable(void);
 
 int copy_file(const char *srcName, const char *destName,
 		 int setModes, int followLinks, int forceFlag);
-int copy_file_chunk(int srcFd, int dstFd, size_t remaining);
+int copy_file_chunk(int srcFd, int dstFd, off_t remaining);
 char *buildName(const char *dirName, const char *fileName);
 int makeString(int argc, const char **argv, char *buf, int bufLen);
 char *getChunk(int size);
